requirements.txt
flask[async]==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
permit==2.7.1
werkzeug==2.3.7
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.1.0
jinja2==3.1.2
pytest==7.4.3
pytest-cov==4.1.0
asgiref==3.7.2